Add Sunmi V2
When printing in the column mode, V2’s printer misinterprets
newline characters and advances the paper too much, making gaps
between bands. This can be avoided by not issuing newlines until
the image has been printed, but better not use this mode at all.

The printer uses monospace fonts, but some characters are double
or 1½ width, like block graphics, €, £ and ₧. This cannot be at
the moment captured in the capabilities file.